+++ src/app/app.module.ts
@@ -6,17 +6,29 @@ import { AppComponent } from './app.component';
 import { HomeComponent } from './pages/home/home.component';
 import { HeaderComponent } from './components/common/header/header.component';
 import { FooterComponent } from './components/common/footer/footer.component';
+import { SearchComponent } from './pages/search/search.component';
+import { MovieDetailsComponent } from './pages/movie-details/movie-details.component';
+import { BannerComponent } from './pages/home/components/banner/banner.component';
+import { ImageComponent } from './components/image/image.component';
+import {HttpClientModule} from "@angular/common/http";
+import { MoviesComponent } from './pages/home/components/movies/movies.component';
 
 @NgModule({
   declarations: [
     AppComponent,
     HomeComponent,
     HeaderComponent,
-    FooterComponent
+    FooterComponent,
+    SearchComponent,
+    MovieDetailsComponent,
+    BannerComponent,
+    ImageComponent,
+    MoviesComponent
   ],
   imports: [
     BrowserModule,
-    AppRoutingModule
+    AppRoutingModule,
+    HttpClientModule
   ],
   providers: [],
   bootstrap: [AppComponent]