+++ src/app/config/API.ts
@@ -1,0 +1,3 @@
+export const BASE_URL: string = "https://api.themoviedb.org/3";
+export const IMAGE_BASE_URL: string = "https://image.tmdb.org/t/p/original";
+export const TRAILER_BASE_URL: string = "https://themoviedb.org/video/play";